refactor(web): ♻️ extract metadata to another file
